Move ArbInfo into preinstalled_contracts.rs
